fix proto
