--- client/crates/types/src/proto/rollkit.v1.services.rs
@@ -1,613 +1,0 @@
-// This file is @generated by prost-build.
-/// InitChainRequest contains the genesis parameters for chain initialization
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct InitChainRequest {
-    /// Timestamp marking chain start time in UTC
-    #[prost(message, optional, tag = "1")]
-    pub genesis_time: ::core::option::Option<::prost_types::Timestamp>,
-    /// First block height (must be > 0)
-    #[prost(uint64, tag = "2")]
-    pub initial_height: u64,
-    /// Unique identifier string for the blockchain
-    #[prost(string, tag = "3")]
-    pub chain_id: ::prost::alloc::string::String,
-}
-/// InitChainResponse contains the initial state and configuration
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct InitChainResponse {
-    /// Hash representing initial state
-    #[prost(bytes = "vec", tag = "1")]
-    pub state_root: ::prost::alloc::vec::Vec<u8>,
-    /// Maximum allowed bytes for transactions in a block
-    #[prost(uint64, tag = "2")]
-    pub max_bytes: u64,
-}
-/// GetTxsRequest is the request for fetching transactions
-///
-/// Empty for now, may include filtering criteria in the future
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GetTxsRequest {}
-/// GetTxsResponse contains the available transactions
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GetTxsResponse {
-    /// Slice of valid transactions from mempool
-    #[prost(bytes = "vec", repeated, tag = "1")]
-    pub txs: ::prost::alloc::vec::Vec<::prost::alloc::vec::Vec<u8>>,
-}
-/// ExecuteTxsRequest contains transactions and block context for execution
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ExecuteTxsRequest {
-    /// Ordered list of transactions to execute
-    #[prost(bytes = "vec", repeated, tag = "1")]
-    pub txs: ::prost::alloc::vec::Vec<::prost::alloc::vec::Vec<u8>>,
-    /// Height of block being created (must be > 0)
-    #[prost(uint64, tag = "2")]
-    pub block_height: u64,
-    /// Block creation time in UTC
-    #[prost(message, optional, tag = "3")]
-    pub timestamp: ::core::option::Option<::prost_types::Timestamp>,
-    /// Previous block's state root hash
-    #[prost(bytes = "vec", tag = "4")]
-    pub prev_state_root: ::prost::alloc::vec::Vec<u8>,
-}
-/// ExecuteTxsResponse contains the result of transaction execution
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ExecuteTxsResponse {
-    /// New state root after executing transactions
-    #[prost(bytes = "vec", tag = "1")]
-    pub updated_state_root: ::prost::alloc::vec::Vec<u8>,
-    /// Maximum allowed transaction size (may change with protocol updates)
-    #[prost(uint64, tag = "2")]
-    pub max_bytes: u64,
-}
-/// SetFinalRequest marks a block as finalized
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct SetFinalRequest {
-    /// Height of block to finalize
-    #[prost(uint64, tag = "1")]
-    pub block_height: u64,
-}
-/// SetFinalResponse indicates whether finalization was successful
-///
-/// Empty response, errors are returned via gRPC status
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct SetFinalResponse {}
-/// Generated client implementations.
-pub mod executor_service_client {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    use tonic::codegen::http::Uri;
-    /// ExecutorService defines the execution layer interface for Rollkit
-    #[derive(Debug, Clone)]
-    pub struct ExecutorServiceClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl ExecutorServiceClient<tonic::transport::Channel> {
-        /// Attempt to create a new client by connecting to a given endpoint.
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> ExecutorServiceClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::Error: Into<StdError>,
-        T::ResponseBody: Body<Data = Bytes> + Send + 'static,
-        <T::ResponseBody as Body>::Error: Into<StdError> + Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_origin(inner: T, origin: Uri) -> Self {
-            let inner = tonic::client::Grpc::with_origin(inner, origin);
-            Self { inner }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> ExecutorServiceClient<InterceptedService<T, F>>
-        where
-            F: tonic::service::Interceptor,
-            T::ResponseBody: Default,
-            T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
-                >,
-            >,
-            <T as tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-            >>::Error: Into<StdError> + Send + Sync,
-        {
-            ExecutorServiceClient::new(InterceptedService::new(inner, interceptor))
-        }
-        /// Compress requests with the given encoding.
-        ///
-        /// This requires the server to support it otherwise it might respond with an
-        /// error.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.send_compressed(encoding);
-            self
-        }
-        /// Enable decompressing responses.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.accept_compressed(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_decoding_message_size(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_encoding_message_size(limit);
-            self
-        }
-        /// InitChain initializes a new blockchain instance with genesis parameters
-        pub async fn init_chain(
-            &mut self,
-            request: impl tonic::IntoRequest<super::InitChainRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::InitChainResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/rollkit.v1.ExecutorService/InitChain",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("rollkit.v1.ExecutorService", "InitChain"));
-            self.inner.unary(req, path, codec).await
-        }
-        /// GetTxs fetches available transactions from the execution layer's mempool
-        pub async fn get_txs(
-            &mut self,
-            request: impl tonic::IntoRequest<super::GetTxsRequest>,
-        ) -> std::result::Result<tonic::Response<super::GetTxsResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/rollkit.v1.ExecutorService/GetTxs",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("rollkit.v1.ExecutorService", "GetTxs"));
-            self.inner.unary(req, path, codec).await
-        }
-        /// ExecuteTxs processes transactions to produce a new block state
-        pub async fn execute_txs(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ExecuteTxsRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::ExecuteTxsResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/rollkit.v1.ExecutorService/ExecuteTxs",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("rollkit.v1.ExecutorService", "ExecuteTxs"));
-            self.inner.unary(req, path, codec).await
-        }
-        /// SetFinal marks a block as finalized at the specified height
-        pub async fn set_final(
-            &mut self,
-            request: impl tonic::IntoRequest<super::SetFinalRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::SetFinalResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/rollkit.v1.ExecutorService/SetFinal",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("rollkit.v1.ExecutorService", "SetFinal"));
-            self.inner.unary(req, path, codec).await
-        }
-    }
-}
-/// Generated server implementations.
-pub mod executor_service_server {
-    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
-    use tonic::codegen::*;
-    /// Generated trait containing gRPC methods that should be implemented for use with ExecutorServiceServer.
-    #[async_trait]
-    pub trait ExecutorService: Send + Sync + 'static {
-        /// InitChain initializes a new blockchain instance with genesis parameters
-        async fn init_chain(
-            &self,
-            request: tonic::Request<super::InitChainRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::InitChainResponse>,
-            tonic::Status,
-        >;
-        /// GetTxs fetches available transactions from the execution layer's mempool
-        async fn get_txs(
-            &self,
-            request: tonic::Request<super::GetTxsRequest>,
-        ) -> std::result::Result<tonic::Response<super::GetTxsResponse>, tonic::Status>;
-        /// ExecuteTxs processes transactions to produce a new block state
-        async fn execute_txs(
-            &self,
-            request: tonic::Request<super::ExecuteTxsRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::ExecuteTxsResponse>,
-            tonic::Status,
-        >;
-        /// SetFinal marks a block as finalized at the specified height
-        async fn set_final(
-            &self,
-            request: tonic::Request<super::SetFinalRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::SetFinalResponse>,
-            tonic::Status,
-        >;
-    }
-    /// ExecutorService defines the execution layer interface for Rollkit
-    #[derive(Debug)]
-    pub struct ExecutorServiceServer<T: ExecutorService> {
-        inner: _Inner<T>,
-        accept_compression_encodings: EnabledCompressionEncodings,
-        send_compression_encodings: EnabledCompressionEncodings,
-        max_decoding_message_size: Option<usize>,
-        max_encoding_message_size: Option<usize>,
-    }
-    struct _Inner<T>(Arc<T>);
-    impl<T: ExecutorService> ExecutorServiceServer<T> {
-        pub fn new(inner: T) -> Self {
-            Self::from_arc(Arc::new(inner))
-        }
-        pub fn from_arc(inner: Arc<T>) -> Self {
-            let inner = _Inner(inner);
-            Self {
-                inner,
-                accept_compression_encodings: Default::default(),
-                send_compression_encodings: Default::default(),
-                max_decoding_message_size: None,
-                max_encoding_message_size: None,
-            }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> InterceptedService<Self, F>
-        where
-            F: tonic::service::Interceptor,
-        {
-            InterceptedService::new(Self::new(inner), interceptor)
-        }
-        /// Enable decompressing requests with the given encoding.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.accept_compression_encodings.enable(encoding);
-            self
-        }
-        /// Compress responses with the given encoding, if the client supports it.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.send_compression_encodings.enable(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.max_decoding_message_size = Some(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.max_encoding_message_size = Some(limit);
-            self
-        }
-    }
-    impl<T, B> tonic::codegen::Service<http::Request<B>> for ExecutorServiceServer<T>
-    where
-        T: ExecutorService,
-        B: Body + Send + 'static,
-        B::Error: Into<StdError> + Send + 'static,
-    {
-        type Response = http::Response<tonic::body::BoxBody>;
-        type Error = std::convert::Infallible;
-        type Future = BoxFuture<Self::Response, Self::Error>;
-        fn poll_ready(
-            &mut self,
-            _cx: &mut Context<'_>,
-        ) -> Poll<std::result::Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-        fn call(&mut self, req: http::Request<B>) -> Self::Future {
-            let inner = self.inner.clone();
-            match req.uri().path() {
-                "/rollkit.v1.ExecutorService/InitChain" => {
-                    #[allow(non_camel_case_types)]
-                    struct InitChainSvc<T: ExecutorService>(pub Arc<T>);
-                    impl<
-                        T: ExecutorService,
-                    > tonic::server::UnaryService<super::InitChainRequest>
-                    for InitChainSvc<T> {
-                        type Response = super::InitChainResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::InitChainRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ExecutorService>::init_chain(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = InitChainSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/rollkit.v1.ExecutorService/GetTxs" => {
-                    #[allow(non_camel_case_types)]
-                    struct GetTxsSvc<T: ExecutorService>(pub Arc<T>);
-                    impl<
-                        T: ExecutorService,
-                    > tonic::server::UnaryService<super::GetTxsRequest>
-                    for GetTxsSvc<T> {
-                        type Response = super::GetTxsResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::GetTxsRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ExecutorService>::get_txs(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = GetTxsSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/rollkit.v1.ExecutorService/ExecuteTxs" => {
-                    #[allow(non_camel_case_types)]
-                    struct ExecuteTxsSvc<T: ExecutorService>(pub Arc<T>);
-                    impl<
-                        T: ExecutorService,
-                    > tonic::server::UnaryService<super::ExecuteTxsRequest>
-                    for ExecuteTxsSvc<T> {
-                        type Response = super::ExecuteTxsResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ExecuteTxsRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ExecutorService>::execute_txs(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = ExecuteTxsSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/rollkit.v1.ExecutorService/SetFinal" => {
-                    #[allow(non_camel_case_types)]
-                    struct SetFinalSvc<T: ExecutorService>(pub Arc<T>);
-                    impl<
-                        T: ExecutorService,
-                    > tonic::server::UnaryService<super::SetFinalRequest>
-                    for SetFinalSvc<T> {
-                        type Response = super::SetFinalResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::SetFinalRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as ExecutorService>::set_final(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = SetFinalSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                _ => {
-                    Box::pin(async move {
-                        Ok(
-                            http::Response::builder()
-                                .status(200)
-                                .header("grpc-status", "12")
-                                .header("content-type", "application/grpc")
-                                .body(empty_body())
-                                .unwrap(),
-                        )
-                    })
-                }
-            }
-        }
-    }
-    impl<T: ExecutorService> Clone for ExecutorServiceServer<T> {
-        fn clone(&self) -> Self {
-            let inner = self.inner.clone();
-            Self {
-                inner,
-                accept_compression_encodings: self.accept_compression_encodings,
-                send_compression_encodings: self.send_compression_encodings,
-                max_decoding_message_size: self.max_decoding_message_size,
-                max_encoding_message_size: self.max_encoding_message_size,
-            }
-        }
-    }
-    impl<T: ExecutorService> Clone for _Inner<T> {
-        fn clone(&self) -> Self {
-            Self(Arc::clone(&self.0))
-        }
-    }
-    impl<T: std::fmt::Debug> std::fmt::Debug for _Inner<T> {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{:?}", self.0)
-        }
-    }
-    impl<T: ExecutorService> tonic::server::NamedService for ExecutorServiceServer<T> {
-        const NAME: &'static str = "rollkit.v1.ExecutorService";
-    }
-}
